move bobkey to vec<u8> usage

+++ src/core/mapper.rs
@@ -120,7 +120,7 @@ impl Virtual {
     }
 
     pub(crate) fn vdisk_id_from_key(&self, key: BobKey) -> VDiskID {
-        (key % self.vdisks.len() as u64)
+        (key % self.vdisks.len())
             .try_into()
             .expect("u64 to u32")
     }
